uniform test fixtures
