+++ bag_register/src/test/scala/uk/ac/wellcome/platform/archive/bag_register/BagRegisterFeatureTest.scala
@@ -5,19 +5,11 @@ import java.time.Instant
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.json.JsonUtil._
 import uk.ac.wellcome.messaging.fixtures.SQS
-import uk.ac.wellcome.platform.archive.bag_register.fixtures.BagRegisterWorkerFixture
+import uk.ac.wellcome.platform.archive.bag_register.fixtures.BagRegisterFixtures
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
 import uk.ac.wellcome.platform.archive.common.fixtures.BagLocationFixtures
-import uk.ac.wellcome.platform.archive.common.generators.{
-  BagInfoGenerators,
-  BagRequestGenerators,
-  IngestOperationGenerators
-}
-import uk.ac.wellcome.platform.archive.common.ingests.models.{
-  InfrequentAccessStorageProvider,
-  Ingest,
-  StorageLocation
-}
+import uk.ac.wellcome.platform.archive.common.generators.{BagInfoGenerators, BagRequestGenerators, IngestOperationGenerators}
+import uk.ac.wellcome.platform.archive.common.ingests.models.{InfrequentAccessStorageProvider, Ingest, StorageLocation}
 import uk.ac.wellcome.platform.archive.common.ingests.fixtures.IngestUpdateAssertions
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
 
@@ -30,7 +22,7 @@ class BagRegisterFeatureTest
     with BagLocationFixtures
     with BagRequestGenerators
     with IngestUpdateAssertions
-    with BagRegisterWorkerFixture {
+    with BagRegisterFixtures {
 
   it("sends an update if it registers a bag") {
     withBagRegisterWorker {
