+++ bag_register/src/test/scala/uk/ac/wellcome/platform/archive/bag_register/services/BagRegisterWorkerTest.scala
@@ -4,19 +4,12 @@ import java.time.Instant
 
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.{FunSpec, Matchers}
-import uk.ac.wellcome.platform.archive.bag_register.fixtures.BagRegisterWorkerFixture
+import uk.ac.wellcome.platform.archive.bag_register.fixtures.BagRegisterFixtures
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagId
 import uk.ac.wellcome.platform.archive.common.fixtures.BagLocationFixtures
-import uk.ac.wellcome.platform.archive.common.generators.{
-  BagInfoGenerators,
-  BagRequestGenerators
-}
-import uk.ac.wellcome.platform.archive.common.ingests.models.{
-  InfrequentAccessStorageProvider,
-  Ingest,
-  StorageLocation
-}
+import uk.ac.wellcome.platform.archive.common.generators.{BagInfoGenerators, BagRequestGenerators}
 import uk.ac.wellcome.platform.archive.common.ingests.fixtures.IngestUpdateAssertions
+import uk.ac.wellcome.platform.archive.common.ingests.models.{InfrequentAccessStorageProvider, Ingest, StorageLocation}
 import uk.ac.wellcome.storage.fixtures.S3.Bucket
 
 class BagRegisterWorkerTest
@@ -27,7 +20,7 @@ class BagRegisterWorkerTest
     with BagLocationFixtures
     with BagRequestGenerators
     with IngestUpdateAssertions
-    with BagRegisterWorkerFixture {
+    with BagRegisterFixtures {
 
   it("sends a successful IngestUpdate upon registration") {
     withBagRegisterWorker {
