+++ bag_replicator/src/test/scala/uk/ac/wellcome/platform/archive/bagreplicator/fixtures/BagReplicatorFixtures.scala
@@ -3,22 +3,71 @@ package uk.ac.wellcome.platform.archive.bagreplicator.fixtures
 import com.amazonaws.services.s3.model.S3ObjectSummary
 import org.scalatest.Assertion
 import uk.ac.wellcome.akka.fixtures.Akka
-import uk.ac.wellcome.messaging.fixtures.Messaging
+import uk.ac.wellcome.fixtures.TestWith
+import uk.ac.wellcome.messaging.fixtures.{Messaging, SQS}
+import uk.ac.wellcome.messaging.fixtures.SNS.Topic
+import uk.ac.wellcome.messaging.fixtures.SQS.Queue
+import uk.ac.wellcome.messaging.fixtures.worker.AlpakkaSQSWorkerFixtures
+import uk.ac.wellcome.messaging.sqsworker.alpakka.AlpakkaSQSWorkerConfig
+import uk.ac.wellcome.platform.archive.bagreplicator.config.ReplicatorDestinationConfig
+import uk.ac.wellcome.platform.archive.bagreplicator.services.{BagReplicator, BagReplicatorWorker}
 import uk.ac.wellcome.platform.archive.common.bagit.models.BagLocation
-import uk.ac.wellcome.platform.archive.common.fixtures.{
-  BagLocationFixtures,
-  RandomThings
-}
+import uk.ac.wellcome.platform.archive.common.fixtures.{BagLocationFixtures, MonitoringClientFixture, OperationFixtures, RandomThings}
 import uk.ac.wellcome.storage.fixtures.S3
+import uk.ac.wellcome.storage.fixtures.S3.Bucket
 
 import scala.collection.JavaConverters._
 
+import scala.concurrent.ExecutionContext.Implicits.global
+
 trait BagReplicatorFixtures
     extends S3
-    with RandomThings
-    with Messaging
-    with Akka
-    with BagLocationFixtures {
+      with SQS
+      with Akka
+      with RandomThings
+      with Messaging
+      with BagLocationFixtures
+      with OperationFixtures
+      with AlpakkaSQSWorkerFixtures
+      with MonitoringClientFixture {
+
+  def withBagReplicatorWorker[R](queue: Queue = Queue(
+    "default_q",
+    "arn::default_q"
+  ),
+                                 ingestTopic: Topic,
+                                 outgoingTopic: Topic,
+                                 config: ReplicatorDestinationConfig =
+                                 createReplicatorDestinationConfigWith(
+                                   Bucket(randomAlphanumeric())))(
+                                  testWith: TestWith[BagReplicatorWorker, R]): R =
+    withActorSystem { implicit actorSystem =>
+      withIngestUpdater("replicating", ingestTopic) { ingestUpdater =>
+        withOutgoingPublisher("replicating", outgoingTopic) {
+          outgoingPublisher =>
+            withMonitoringClient { implicit monitoringClient =>
+              val service = new BagReplicatorWorker(
+                alpakkaSQSWorkerConfig =
+                  AlpakkaSQSWorkerConfig("test", queue.url),
+                bagReplicator = new BagReplicator(config),
+                ingestUpdater = ingestUpdater,
+                outgoingPublisher = outgoingPublisher
+              )
+
+              service.run()
+
+              testWith(service)
+            }
+        }
+      }
+    }
+
+  def createReplicatorDestinationConfigWith(
+                                             bucket: Bucket): ReplicatorDestinationConfig =
+    ReplicatorDestinationConfig(
+      namespace = bucket.name,
+      rootPath = Some(randomAlphanumeric())
+    )
 
   def verifyBagCopied(src: BagLocation, dst: BagLocation): Assertion = {
     val sourceItems = getObjectSummaries(src)
