+++ bag_unpacker/src/test/scala/uk/ac/wellcome/platform/archive/bagunpacker/fixtures/BagUnpackerFixtures.scala
@@ -38,7 +38,6 @@ trait BagUnpackerFixtures
       withIngestUpdater("unpacker", ingestTopic) { ingestUpdater =>
         withOutgoingPublisher("unpacker", outgoingTopic) { ongoingPublisher =>
           withMonitoringClient { implicit monitoringClient =>
-            implicit val _asyncSqsClient = asyncSqsClient
             val bagUnpackerWorker = BagUnpackerWorker(
               alpakkaSQSWorkerConfig = AlpakkaSQSWorkerConfig("test", queue.url),
               bagUnpackerWorkerConfig = BagUnpackerWorkerConfig(dstBucket.name),
