+++ ingests/src/test/scala/uk/ac/wellcome/platform/archive/ingests/IngestsFeatureTest.scala
@@ -14,7 +14,7 @@ class IngestsFeatureTest
     extends FunSpec
     with Matchers
     with ScalaFutures
-    with IngestsFixture
+    with IngestsFixtures
     with IntegrationPatience {
 
   it("updates an existing ingest status to Completed") {
