+++ ingests/src/test/scala/uk/ac/wellcome/platform/archive/ingests/services/IngestsWorkerServiceTest.scala
@@ -7,26 +7,19 @@ import uk.ac.wellcome.messaging.fixtures.SNS.Topic
 import uk.ac.wellcome.messaging.worker.models.DeterministicFailure
 import uk.ac.wellcome.platform.archive.common.generators.IngestGenerators
 import uk.ac.wellcome.platform.archive.common.ingests.models.CallbackNotification
-import uk.ac.wellcome.platform.archive.common.ingests.models.Ingest.{
-  Completed,
-  Processing
-}
+import uk.ac.wellcome.platform.archive.common.ingests.models.Ingest.{Completed, Processing}
 import uk.ac.wellcome.platform.archive.common.ingests.monitor.IdConstraintError
-import uk.ac.wellcome.platform.archive.ingests.fixtures.{
-  IngestsFixture,
-  WorkerServiceFixture
-}
+import uk.ac.wellcome.platform.archive.ingests.fixtures.IngestsFixtures
 
 class IngestsWorkerServiceTest
     extends FunSpec
     with IngestGenerators
-    with IngestsFixture
-    with WorkerServiceFixture {
+    with IngestsFixtures {
   it("updates an existing ingest to Completed") {
     withLocalSqsQueue { queue =>
       withIngestTrackerTable { table =>
         withLocalSnsTopic { topic =>
-          withWorkerService(queue, table, topic) { service =>
+          withIngestWorker(queue, table, topic) { service =>
             withIngestTracker(table) { monitor =>
               withIngest(monitor) { ingest =>
                 val ingestStatusUpdate =
@@ -75,7 +68,7 @@ class IngestsWorkerServiceTest
     withLocalSqsQueue { queue =>
       withIngestTrackerTable { table =>
         withLocalSnsTopic { topic =>
-          withWorkerService(queue, table, topic) { service =>
+          withIngestWorker(queue, table, topic) { service =>
             withIngestTracker(table) { monitor =>
               withIngest(monitor) { ingest =>
                 val ingestStatusUpdate1 =
@@ -126,7 +119,7 @@ class IngestsWorkerServiceTest
     withLocalSqsQueue { queue =>
       withIngestTrackerTable { table =>
         withLocalSnsTopic { topic =>
-          withWorkerService(queue, table, topic) { service =>
+          withIngestWorker(queue, table, topic) { service =>
             val ingestStatusUpdate =
               createIngestStatusUpdateWith(
                 status = Completed
@@ -149,7 +142,7 @@ class IngestsWorkerServiceTest
   it("fails if publishing to SNS fails") {
     withLocalSqsQueue { queue =>
       withIngestTrackerTable { table =>
-        withWorkerService(queue, table, Topic("does-not-exist")) { service =>
+        withIngestWorker(queue, table, Topic("does-not-exist")) { service =>
           withIngestTracker(table) { monitor =>
             withIngest(monitor) { ingest =>
               val ingestStatusUpdate =
