+++ notifier/src/test/scala/uk/ac/wellcome/platform/archive/notifier/fixtures/NotifierFixtures.scala
@@ -1,27 +1,34 @@
 package uk.ac.wellcome.platform.archive.notifier.fixtures
 
+import java.net.URL
+
+import akka.actor.ActorSystem
 import uk.ac.wellcome.fixtures.TestWith
 import uk.ac.wellcome.messaging.fixtures.SNS
 import uk.ac.wellcome.messaging.fixtures.SNS.Topic
 import uk.ac.wellcome.messaging.fixtures.SQS.Queue
 import uk.ac.wellcome.messaging.fixtures.worker.AlpakkaSQSWorkerFixtures
-import uk.ac.wellcome.platform.archive.common.fixtures.{
-  BagIt,
-  MonitoringClientFixture
-}
-import uk.ac.wellcome.platform.archive.notifier.services.NotifierWorker
+import uk.ac.wellcome.platform.archive.common.fixtures.{BagIt, MonitoringClientFixture}
+import uk.ac.wellcome.platform.archive.notifier.services.{CallbackUrlService, NotifierWorker}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
-trait WorkerServiceFixture
-    extends BagIt
-    with CallbackUrlServiceFixture
-    with AlpakkaSQSWorkerFixtures
-    with MonitoringClientFixture
-    with SNS {
+trait NotifierFixtures
+  extends BagIt
+  with AlpakkaSQSWorkerFixtures
+  with MonitoringClientFixture
+  with SNS {
+
+  def withCallbackUrlService[R](testWith: TestWith[CallbackUrlService, R])(
+    implicit actorSystem: ActorSystem): R = {
+    val callbackUrlService = new CallbackUrlService(
+      contextURL = new URL("http://localhost/context.json")
+    )
+    testWith(callbackUrlService)
+  }
 
   private def withApp[R](queue: Queue, topic: Topic)(
-    testWith: TestWith[NotifierWorker, R]): R =
+      testWith: TestWith[NotifierWorker, R]): R =
     withMonitoringClient { implicit monitoringClient =>
       withActorSystem { implicit actorSystem =>
         withMaterializer(actorSystem) { implicit materializer =>
