+++ notifier/src/test/scala/uk/ac/wellcome/platform/archive/notifier/services/CallbackUrlServiceTest.scala
@@ -10,7 +10,7 @@ import uk.ac.wellcome.akka.fixtures.Akka
 import uk.ac.wellcome.platform.archive.common.generators.IngestGenerators
 import uk.ac.wellcome.platform.archive.common.ingests.models.CallbackNotification
 import uk.ac.wellcome.platform.archive.notifier.fixtures.{
-  CallbackUrlServiceFixture,
+  NotifierFixtures,
   LocalWireMockFixture
 }
 
@@ -20,7 +20,7 @@ class CallbackUrlServiceTest
     with ScalaFutures
     with IntegrationPatience
     with Akka
-    with CallbackUrlServiceFixture
+    with NotifierFixtures
     with LocalWireMockFixture
     with IngestGenerators {
